refactor(empty): Moved account row to a separate component so that use client can be moved down
